requirements.txt
asyncpg
Jinja2
openai
pydantic
pydantic-settings
python-telegram-bot
sqlalchemy[asyncio]
